Форматтер

+++ app/db/models/world_setting.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from decimal import Decimal
+
 from sqlalchemy import Numeric, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -10,5 +12,5 @@ class WorldSetting(Base):
     __tablename__ = "world_settings"
 
     name: Mapped[str] = mapped_column(String(255), primary_key=True)
-    value: Mapped[float] = mapped_column(Numeric(10, 4), nullable=False)
+    value: Mapped[Decimal] = mapped_column(Numeric(10, 4), nullable=False)
     description: Mapped[str] = mapped_column(Text, nullable=False)
